Add python-dotenv to the environment and requirements
This commit introduces a new environment file `environment.yaml` for the `kobis-crawl` application and adds `python-dotenv==1.0.0` to the dependencies listed in both `environment.yaml` and `requirements.txt`.

The `python-dotenv` module was added to allow the application to load environment variables from a `.env` file, enhancing the configurability of the application. This is eventually beneficial when dealing with secrets and configurations that should not be hard-coded into the application.