Refactor method to static in DailyBoxOfficeCrawler
The method 'validate_values' in 'DailyBoxOfficeCrawler' does not use or alter any instance states. To better follow the design principle, this method was refactored to a static method in order to clarify that its behavior is independent of the state of the instance. The change will not affect the functionality but enhances the understandability of the code.