Add daily box office posting capability
Introduced a new feature that allows for daily box office results to be posted automatically. This involves crawling Kobis for the box office data, formatting it, and then posting to a Tistory blog.
Also, added additional attributes to the DailyBoxOffice model to allow for more detailed posts including data like daily audience count, cumulative audience count, daily revenue and cumulative revenue.
A Tistory model was also created to handle the requests to the Tistory API for posting to the blog.