Update page layout and dynamic date display
Updated the layout of the daily box office page for better user readability and visual appeal. Also, added a dynamic date display instead of static dates on the box office list. Adaptive margin, custom box shadow, padding, and color scheme were added to the contents for a cleaner look. The date string is now passed as context which makes it dynamic in the templates. These changes aim to enhance user experience and reduce manual updates of dates in the HTML file.